Rename xmvn-connector-aether to xmvn-connector
Closes #93
