+++ doc/install-dfd.dia
@@ -598,7 +598,7 @@ plan#</dia:string>
       <dia:attribute name="text">
         <dia:composite type="text">
           <dia:attribute name="string">
-            <dia:string>#Aether#</dia:string>
+            <dia:string>#Maven Resolver#</dia:string>
           </dia:attribute>
           <dia:attribute name="font">
             <dia:font family="sans" style="0" name="Helvetica"/>
@@ -643,7 +643,7 @@ plan#</dia:string>
       <dia:attribute name="text">
         <dia:composite type="text">
           <dia:attribute name="string">
-            <dia:string>#xmvn-connector-aether#</dia:string>
+            <dia:string>#xmvn-connector#</dia:string>
           </dia:attribute>
           <dia:attribute name="font">
             <dia:font family="sans" style="0" name="Helvetica"/>
