+++ pom.xml
@@ -38,14 +38,14 @@
     <module>xmvn-parent</module>
     <module>xmvn-api</module>
     <module>xmvn-core</module>
-    <module>xmvn-connector-aether</module>
+    <module>xmvn-connector</module>
     <module>xmvn-mojo</module>
     <module>xmvn-tools</module>
   </modules>
   <dependencies>
     <dependency>
       <groupId>org.fedoraproject.xmvn</groupId>
-      <artifactId>xmvn-connector-aether</artifactId>
+      <artifactId>xmvn-connector</artifactId>
       <scope>runtime</scope>
     </dependency>
     <dependency>
