--- xmvn-connector-aether/src/main/java/org/fedoraproject/xmvn/connector/aether/ResolutionListener.java
@@ -1,29 +1,0 @@
-/*-
- * Copyright (c) 2014-2021 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.fedoraproject.xmvn.connector.aether;
-
-import org.fedoraproject.xmvn.resolver.ResolutionRequest;
-import org.fedoraproject.xmvn.resolver.ResolutionResult;
-
-/**
- * @author Mikolaj Izdebski
- */
-public interface ResolutionListener
-{
-    void resolutionRequested( ResolutionRequest request );
-
-    void resolutionCompleted( ResolutionRequest request, ResolutionResult result );
-}
