+++ xmvn-connector/src/test/resources/test.pom.xml
