+++ xmvn-it/src/test/resources/m2.conf
@@ -5,6 +5,6 @@ set maven.home default ${user.home}/m2
 [plexus.core]
 load       ${xmvn.it.rootDir}/xmvn-api/target/classes
 load       ${xmvn.it.rootDir}/xmvn-core/target/classes
-load       ${xmvn.it.rootDir}/xmvn-connector-aether/target/classes
+load       ${xmvn.it.rootDir}/xmvn-connector/target/classes
 load       ${maven.home}/lib/*.jar
 load       ${maven.home}/conf/logging
